refactor: restructure project detailsfragment DI
